+++ common/src/main/java/net/SpectrumFATM/black_archive/blockentity/ModModels.java
@@ -2,8 +2,6 @@ package net.SpectrumFATM.black_archive.blockentity;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.SpectrumFATM.BlackArchive;
-import net.SpectrumFATM.black_archive.blockentity.door.PillarDoorModel;
-import net.SpectrumFATM.black_archive.blockentity.shell.PillarShellModel;
 import net.SpectrumFATM.black_archive.blockentity.shell.ShellEntryRegistry;
 import net.minecraft.client.model.geom.EntityModelSet;
 import net.minecraft.client.model.geom.ModelLayerLocation;
@@ -17,13 +15,13 @@ import java.util.function.Supplier;
 
 public class ModModels {
 
-    public static PillarShellModel PILLAR_SHELL_MODEL;
+    //public static PillarShellModel PILLAR_SHELL_MODEL;
 
-    public static PillarDoorModel PILLAR_DOOR_MODEL;
+    //public static PillarDoorModel PILLAR_DOOR_MODEL;
 
-    public static ModelLayerLocation PILLAR_SHELL = new ModelLayerLocation(new ResourceLocation(BlackArchive.MOD_ID, "pillar_shell"), "shell");
-    public static ModelLayerLocation PILLAR_DOOR = new ModelLayerLocation(new ResourceLocation(BlackArchive.MOD_ID, "pillar_door"), "door");
-    public static ModelLayerLocation RANI_CONSOLE = new ModelLayerLocation(new ResourceLocation(BlackArchive.MOD_ID, "rani_console"), "console");
+    //public static ModelLayerLocation PILLAR_SHELL = new ModelLayerLocation(new ResourceLocation(BlackArchive.MOD_ID, "pillar_shell"), "shell");
+    //public static ModelLayerLocation PILLAR_DOOR = new ModelLayerLocation(new ResourceLocation(BlackArchive.MOD_ID, "pillar_door"), "door");
+    //public static ModelLayerLocation RANI_CONSOLE = new ModelLayerLocation(new ResourceLocation(BlackArchive.MOD_ID, "rani_console"), "console");
 
 
     public static void init() {
@@ -33,8 +31,8 @@ public class ModModels {
     }
 
     public static void setupModelInstances(EntityModelSet entityModels) {
-        ModModels.PILLAR_SHELL_MODEL = new PillarShellModel(entityModels.bakeLayer(ModModels.PILLAR_SHELL));
-        ModModels.PILLAR_DOOR_MODEL = new PillarDoorModel(entityModels.bakeLayer(ModModels.PILLAR_DOOR));
+        //ModModels.PILLAR_SHELL_MODEL = new PillarShellModel(entityModels.bakeLayer(ModModels.PILLAR_SHELL));
+        //ModModels.PILLAR_DOOR_MODEL = new PillarDoorModel(entityModels.bakeLayer(ModModels.PILLAR_DOOR));
         ShellEntryRegistry.init();
     }