+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Sonic screwdriver can now be used to interact with the astral manipulator.
 - Time fissures can now be closed with the interact setting on the sonic screwdriver.
 - The interact setting on the sonic now works on doors and trap doors.
+- Vortex manipulator no longer works in the TARDIS interior.
+- Shell viewing no longer causes damage in space.
 
 ### Added
 - Time fissures now have a chance of spawning Silurians and Weeping Angels, as well as the already added mobs.
@@ -12,6 +14,8 @@
 - Sonic pen.
 - War Doctor's Sonic Screwdriver.
 - Ad Astra compatability.
+- Mondas planet.
 
 ### Removed
 - Rani console, exterior and interior.
+- Pillar exterior.